Task/ssup 29 game end 2

+++ Assets/Game/Modules/Player/Scripts/Ship/PlayerShipInstaller.cs
@@ -1,7 +1,6 @@
 ﻿using Game.Modules.ShootingModule.Scripts;
 using Game.Modules.ShootingModule.Scripts.ScriptableObjects;
 using SpaceShooter.Game.Components;
-using SpaceShooter.Game.LifeCycle.Common;
 using UnityEngine;
 using Zenject;
 
@@ -10,7 +9,7 @@ namespace SpaceShooter.Game.Player.Ship
     public class PlayerShipInstaller : MonoInstaller
     {
         [SerializeField] private float speed;
-        [SerializeField] private float health;
+        [SerializeField] private int health;
         
         [SerializeField] private WeaponConfig weaponConfig;
